A3

+++ kern/arch/mips/vm/dumbvm.c
@@ -147,7 +147,7 @@ paddr_t vm_stealmem(unsigned long npages) {
 
 				for (unsigned long npages_to_allocate = 0; npages_to_allocate < npage_can_allocate; npages_to_allocate++){
 
-					((int *) PADDR_TO_KVADDR(coremap_start))[cur_page + npages_to_allocate] = npages_to_allocate + 1;
+					((int *) PADDR_TO_KVADDR(coremap_start))[cur_page + npages_to_allocate] = (int)(npages_to_allocate + 1);
 					//*(int *)(coremap_start + (npages_to_allocate * sizeof(int))) = npages_to_allocate + 1;
 
 				}
